recon_tool/requirements.txt
flask
requests
dnspython
python-whois
jinja2
